% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asana_api_request}
\alias{asana_api_request}
\title{Send httr requests to Asana API}
\usage{
asana_api_request(verb = c("GET", "POST", "PUT", "DELETE"), url = url,
  body = NULL)
}
\arguments{
\item{verb}{verb type of request to be sent (@seealso ?httr::VERB)}

\item{url}{url Asana endpoint url created by and passed to this function}

\item{body}{body options created by and passed to this function}
}
\description{
Send httr requests to Asana API
}
\author{
Ryan Sullivan
}
