% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asana_errors}
\alias{asana_errors}
\title{Handle Asana API Response Codes}
\usage{
asana_errors(resp, parsed)
}
\arguments{
\item{resp}{A Response object from the Asana API}

\item{parsed}{The Parsed Response}
}
\description{
Handle Asana API Response Codes
}
\author{
Ryan Sullivan
}
