% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asanar-package.R
\docType{package}
\name{asanar-package}
\alias{asanar}
\alias{asanar-package}
\title{asanar: 'Asana' 'API'}
\description{
Client to access the 'Asana' 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/asanar}
  \item \url{https://jonthegeek.github.io/asanar/}
  \item Report bugs at \url{https://github.com/jonthegeek/asanar/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
