% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asanaRconfigFile.R
\name{asanaRconfigFile}
\alias{asanaRconfigFile}
\title{Prints an example of a asanaR configuration file to the console.}
\usage{
asanaRconfigFile(personal_access_token = NULL)
}
\arguments{
\item{personal_access_token}{This is your personal access token from Asana}
}
\description{
Prints an example of a asanaR configuration file to the console.
}
\examples{
\dontrun{
# Execute this line to get instructions on how to make a .asana.yml config file.
asanaRconfigFile()
}

}
\author{
Ryan Sullivan
}
