% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asanaResponseCodes}
\alias{asanaResponseCodes}
\title{Checks responses against Asana response codes and returns an error message.}
\usage{
asanaResponseCodes(res, raw = FALSE)
}
\arguments{
\item{res}{response from httr::GET}

\item{raw}{raw if TRUE, add 'raw' flag to httr::content() function.}
}
\description{
Checks responses against Asana response codes and returns an error message.
}
\author{
Ryan Sullivan
}
