% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{asn_add_custom_field_setting_for_portfolio}
\alias{asn_add_custom_field_setting_for_portfolio}
\title{Add a custom field to a portfolio}
\usage{
asn_add_custom_field_setting_for_portfolio(portfolio_gid)
}
\arguments{
\item{portfolio_gid}{(character scalar) Globally unique identifier for the portfolio.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Custom fields are associated with portfolios by way of custom field settings.  This method creates a setting for the portfolio.
}
\keyword{internal}
