% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{asn_add_followers}
\alias{asn_add_followers}
\title{Add a collaborator to a goal}
\usage{
asn_add_followers(goal_gid, opt_fields)
}
\arguments{
\item{goal_gid}{(character scalar) Globally unique identifier for the goal.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Adds followers to a goal. Returns the goal the followers were added to. Each goal can be associated with zero or more followers in the system. Requests to add/remove followers, if successful, will return the complete updated goal record, described above.
}
\keyword{internal}
