% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_add_members_for_project}
\alias{asn_add_members_for_project}
\title{Add users to a project}
\usage{
asn_add_members_for_project(project_gid, opt_fields)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Adds the specified list of users as members of the project. Note that a user being added as a member may also be added as a \emph{follower} as a result of this operation. This is because the user's default notification settings (i.e., in the "Notifcations" tab of "My Profile Settings") will override this endpoint's default behavior of setting "Tasks added" notifications to \code{false}. Returns the updated project record.
}
\keyword{internal}
