% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_add_project_for_task}
\alias{asn_add_project_for_task}
\title{Add a project to a task}
\usage{
asn_add_project_for_task(task_gid, opt_fields)
}
\arguments{
\item{task_gid}{(character scalar) The task to operate on.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
Adds the task to the specified project, in the optional location specified. If no location arguments are given, the task will be added to the end of the project.  \code{addProject} can also be used to reorder a task within a project or section that already contains it.  At most one of \code{insert_before}, \code{insert_after}, or \code{section} should be specified. Inserting into a section in an non-order-dependent way can be done by specifying section, otherwise, to insert within a section in a particular place, specify \code{insert_before} or \code{insert_after} and a task within the section to anchor the position of this task.  Returns an empty data block.
}
\keyword{internal}
