% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_add_task_for_section}
\alias{asn_add_task_for_section}
\title{Add task to section}
\usage{
asn_add_task_for_section(section_gid, opt_fields)
}
\arguments{
\item{section_gid}{(character scalar) The globally unique identifier for the section.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
Add a task to a specific, existing section. This will remove the task from other sections of the project.  The task will be inserted at the top of a section unless an insert_before or insert_after parameter is declared.  This does not work for separators (tasks with the resource_subtype of section).
}
\keyword{internal}
