% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_add_user_for_team}
\alias{asn_add_user_for_team}
\title{Add a user to a team}
\usage{
asn_add_user_for_team(team_gid, opt_fields)
}
\arguments{
\item{team_gid}{(character scalar) Globally unique identifier for the team.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
The user making this call must be a member of the team in order to add others. The user being added must exist in the same organization as the team.  Returns the complete team membership record for the newly added user.
}
\keyword{internal}
