% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_authentication.R
\name{asn_browse_token}
\alias{asn_browse_token}
\title{Visit Asana My Apps page}
\usage{
asn_browse_token()
}
\value{
The url of the "My Apps" page, invisibly.
}
\description{
Launch the Asana "My Apps" page (where you can create a new personal access
token). After you create a token, copy/paste the value into the
\code{ASN_API_TOKEN} environment variable. We recommend placing this environment
variable in your \code{.Renviron} file.
}
\examples{
asana_token_url <- asn_browse_token()
Sys.setenv(ASN_API_TOKEN = "1/12345:e12a3etc")
}
