% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{asn_create_batch_request}
\alias{asn_create_batch_request}
\title{Submit parallel requests}
\usage{
asn_create_batch_request(opt_fields)
}
\arguments{
\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A response object returned from a batch request.
}
\description{
Make multiple requests in parallel to Asana's API.
}
\keyword{internal}
