% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{asn_create_goal_metric}
\alias{asn_create_goal_metric}
\title{Create a goal metric}
\usage{
asn_create_goal_metric(goal_gid, opt_fields)
}
\arguments{
\item{goal_gid}{(character scalar) Globally unique identifier for the goal.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creates and adds a goal metric to a specified goal. Note that this replaces an existing goal metric if one already exists.
}
\keyword{internal}
