% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{asn_create_membership}
\alias{asn_create_membership}
\title{Create a membership}
\usage{
asn_create_membership(opt_fields)
}
\arguments{
\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creates a new membership in a \code{team}, \code{project}, \code{goal}, or \code{portfolio}. \code{Teams} or \code{users} can be a member of \code{goals}. \code{Project}, \code{team}, and \code{portfolios} have users as members.  Returns the full record of the newly created membership.
}
\keyword{internal}
