% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{asn_create_portfolio}
\alias{asn_create_portfolio}
\title{Create a portfolio}
\usage{
asn_create_portfolio(opt_fields)
}
\arguments{
\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creates a new portfolio in the given workspace with the supplied name.  Note that portfolios created in the Asana UI may have some state (like the “Priority” custom field) which is automatically added to the portfolio when it is created. Portfolios created via our API will \emph{not} be created with the same initial state to allow integrations to create their own starting state on a portfolio.
}
\keyword{internal}
