% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_create_project}
\alias{asn_create_project}
\title{Create a project}
\usage{
asn_create_project(opt_fields)
}
\arguments{
\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Create a new project in a workspace or team.  Every project is required to be created in a specific workspace or organization, and this cannot be changed once set. Note that you can use the \code{workspace} parameter regardless of whether or not it is an organization.  If the workspace for your project is an organization, you must also supply a \code{team} to share the project with.  Returns the full record of the newly created project.
}
\keyword{internal}
