% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_create_project_brief}
\alias{asn_create_project_brief}
\title{Create a project brief}
\usage{
asn_create_project_brief(project_gid, opt_fields)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A \emph{Project Brief} allows you to explain the what and why of the project to your team.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   html_text \tab character scalar \tab HTML formatted text for the project brief. \cr
   title \tab character scalar \tab The title of the project brief. \cr
   permalink_url \tab character scalar \tab A url that points directly to the object within Asana. \cr
   project \tab list \tab The project with which this project brief is associated. \cr
   text \tab character scalar \tab \href{/docs/inputoutput-options}{Opt In}. The plain text of the project brief. \cr
}
}
\description{
Creates a new project brief.  Returns the full record of the newly created project brief.
}
\keyword{internal}
