% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_create_project_for_team}
\alias{asn_create_project_for_team}
\title{Create a project in a team}
\usage{
asn_create_project_for_team(team_gid, opt_fields)
}
\arguments{
\item{team_gid}{(character scalar) Globally unique identifier for the team.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creates a project shared with the given team.  Returns the full record of the newly created project.
}
\keyword{internal}
