% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_create_project_status_for_project}
\alias{asn_create_project_status_for_project}
\title{Create a project status}
\usage{
asn_create_project_status_for_project(project_gid, opt_fields)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
\emph{Deprecated: new integrations should prefer the \code{status_update} resource.} A \emph{project status} is an update on the progress of a particular project, and is sent out to all project followers when created. These updates include both text describing the update and a color code intended to represent the overall state of the project: "green" for projects that are on track, "yellow" for projects at risk, and "red" for projects that are behind.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   title \tab character scalar \tab The title of the project status update. \cr
   color \tab character scalar \tab The color associated with the status update. \cr
   html_text \tab character scalar \tab \href{/docs/inputoutput-options}{Opt In}. The text content of the status update with formatting as HTML. \cr
   text \tab character scalar \tab The text content of the status update. \cr
   author \tab list \tab A \emph{user} object represents an account in Asana that can be given access to various workspaces, projects, and tasks. \cr
   created_at \tab datetime scalar \tab The time at which this resource was created. \cr
   created_by \tab list \tab A \emph{user} object represents an account in Asana that can be given access to various workspaces, projects, and tasks. \cr
   modified_at \tab datetime scalar \tab The time at which this project status was last modified. \emph{Note: This does not currently reflect any changes in associations such as comments that may have been added or removed from the project status.} \cr
}
}
\description{
\emph{Deprecated: new integrations should prefer the \verb{/status_updates} route.}  Creates a new status update on the project.  Returns the full record of the newly created project status update.
}
\keyword{internal}
