% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_create_story_for_task}
\alias{asn_create_story_for_task}
\title{Create a story on a task}
\usage{
asn_create_story_for_task(task_gid, opt_fields)
}
\arguments{
\item{task_gid}{(character scalar) The task to operate on.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Adds a story to a task. This endpoint currently only allows for comment stories to be created. The comment will be authored by the currently authenticated user, and timestamped when the server receives the request.  Returns the full record for the new story added to the task.
}
\keyword{internal}
