% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_create_task}
\alias{asn_create_task}
\title{Create a task}
\usage{
asn_create_task(opt_fields)
}
\arguments{
\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creating a new task is as easy as POSTing to the \verb{/tasks} endpoint with a data block containing the fields you’d like to set on the task. Any unspecified fields will take on default values.  Every task is required to be created in a specific workspace, and this workspace cannot be changed once set. The workspace need not be set explicitly if you specify \code{projects} or a \code{parent} task instead.
}
\keyword{internal}
