% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_statuses.R
\name{asn_delete_project_status}
\alias{asn_delete_project_status}
\title{Delete a project status}
\usage{
asn_delete_project_status(project_status_gid, opt_fields)
}
\arguments{
\item{project_status_gid}{(character scalar) The project status update to get.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
\emph{Deprecated: new integrations should prefer the \verb{/status_updates/\{status_gid\}} route.}  Deletes a specific, existing project status update.  Returns an empty data record.
}
\keyword{internal}
