% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_delete_section}
\alias{asn_delete_section}
\title{Delete a section}
\usage{
asn_delete_section(section_gid, opt_fields)
}
\arguments{
\item{section_gid}{(character scalar) The globally unique identifier for the section.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
A specific, existing section can be deleted by making a DELETE request on the URL for that section.  Note that sections must be empty to be deleted.  The last remaining section cannot be deleted.  Returns an empty data block.
}
\keyword{internal}
