% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{asn_delete_tag}
\alias{asn_delete_tag}
\title{Delete a tag}
\usage{
asn_delete_tag(tag_gid, limit, offset, opt_fields)
}
\arguments{
\item{tag_gid}{(character scalar) Globally unique identifier for the tag.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
A specific, existing tag can be deleted by making a DELETE request on the URL for that tag.  Returns an empty data record.
}
\keyword{internal}
