% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{asn_delete_webhook}
\alias{asn_delete_webhook}
\title{Delete a webhook}
\usage{
asn_delete_webhook(webhook_gid, opt_fields)
}
\arguments{
\item{webhook_gid}{(character scalar) Globally unique identifier for the webhook.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
This method \emph{permanently} removes a webhook. Note that it may be possible to receive a request that was already in flight after deleting the webhook, but no further requests will be issued.
}
\keyword{internal}
