% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{asn_get_events}
\alias{asn_get_events}
\title{Get events on a resource}
\usage{
asn_get_events(resource, opt_fields, sync)
}
\arguments{
\item{resource}{(character scalar) A resource ID to subscribe to. The resource can be a task or project.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{sync}{(optional) (character scalar) A sync token received from the last request, or none on first sync. Events will be returned from the point in time that the sync token was generated. \emph{Note: On your first request, omit the sync token. The response will be the same as for an expired sync token, and will include a new valid sync token.If the sync token is too old (which may happen from time to time) the API will return a \verb{412 Precondition Failed} error, and include a fresh sync token in the response.}}
}
\value{
An \emph{event} is an object representing a change to a resource that was observed by an event subscription or delivered asynchronously to the target location of an active webhook.  The event may be triggered by a different \code{user} than the subscriber. For example, if user A subscribes to a task and user B modified it, the event’s user will be user B. Note: Some events are generated by the system, and will have \code{null} as the user. API consumers should make sure to handle this case.  The \code{resource} that triggered the event may be different from the one that the events were requested for or the webhook is subscribed to. For example, a subscription to a project will contain events for tasks contained within the project.  \strong{Note:} pay close attention to the relationship between the fields \code{Event.action} and \code{Event.change.action}. \code{Event.action} represents the action taken on the resource itself, and \code{Event.change.action} represents how the information within the resource's fields have been modified.  For instance, consider these scenarios:   * When at task is added to a project, \code{Event.action} will be \code{added}, \code{Event.parent} will be an object with the \code{id} and \code{type} of the project, and there will be no \code{change} field.   * When an assignee is set on the task, \code{Event.parent} will be \code{null}, \code{Event.action} will be \code{changed}, \code{Event.change.action} will be \code{changed}, and \code{new_value} will be an object with the user's \code{id} and \code{type}.   * When a collaborator is added to the task, \code{Event.parent} will be \code{null}, \code{Event.action} will be \code{changed}, \code{Event.change.action} will be \code{added}, and \code{added_value} will be an object with the user's \code{id} and \code{type}.
}
\description{
Returns the full record for all events that have occurred since the sync token was created.  A \code{GET} request to the endpoint \verb{/[path_to_resource]/events} can be made in lieu of including the resource ID in the data for the request.  Asana limits a single sync token to 100 events. If more than 100 events exist for a given resource, \code{has_more: true} will be returned in the response, indicating that there are more events to pull.   \emph{Note: The resource returned will be the resource that triggered the event. This may be different from the one that the events were requested for. For example, a subscription to a project will contain events for tasks contained within the project.}
}
\keyword{internal}
