% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{asn_get_favorites_for_user}
\alias{asn_get_favorites_for_user}
\title{Get a user's favorites}
\usage{
asn_get_favorites_for_user(
  resource_type = "project",
  user_gid,
  workspace,
  opt_fields
)
}
\arguments{
\item{resource_type}{(character scalar) The resource type of favorites to be returned. Possible values: portfolio, project, tag, task, user, project_template. Default: project}

\item{user_gid}{(character scalar) A string identifying a user. This can either be the string "me", an email, or the gid of a user.}

\item{workspace}{(character scalar) The workspace in which to get favorites.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab The name of the object. \cr
}
}
\description{
Returns all of a user's favorites in the given workspace, of the given type. Results are given in order (The same order as Asana's sidebar).
}
\keyword{internal}
