% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{asn_get_memberships}
\alias{asn_get_memberships}
\title{Get multiple memberships}
\usage{
asn_get_memberships(
  limit,
  member,
  offset,
  opt_fields,
  parent,
  resource_subtype,
  workspace
)
}
\arguments{
\item{limit}{(optional) (integer scalar) Pagination limit for the request.}

\item{member}{(optional) (character scalar) Globally unique identifier for \code{team} or \code{user}.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{parent}{(optional) (character scalar) Globally unique identifier for \code{project}, \code{portfolio},   \code{team}, \code{goal}, and \code{workspace}.}

\item{resource_subtype}{(optional) (character scalar) The resource_subtype to filter on. Must be provided with \code{member} and \code{workspace} if \code{parent} is not provided. Valid values include \code{team_membership}, \code{workspace_membership}, \code{portfolio_membership}}

\item{workspace}{(optional) (character scalar) The workspace to filter on. Must be provided with \code{member} and \code{resource_subtype} if \code{parent} is not provided.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns compact \code{goal_membership}, \code{team_membership}, \code{project_membership}, \code{portfolio_membership}, or \code{workspace_membership} records. The possible types for \code{parent} in this request are \code{project}, \code{portfolio}, \code{team}, \code{goal}, and \code{workspace}. An additional member (user GID or team GID) can be passed in to filter to a specific membership. If a \code{parent} param is not provided, a \code{member}, \code{resource_subtype}, and \code{workspace} param must be provided.
}
\keyword{internal}
