% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\name{asn_get_parent_goals_for_goal}
\alias{asn_get_parent_goals_for_goal}
\title{Get parent goals from a goal}
\usage{
asn_get_parent_goals_for_goal(goal_gid, opt_fields)
}
\arguments{
\item{goal_gid}{(character scalar) Globally unique identifier for the goal.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab The name of the goal. \cr
   owner \tab list \tab A \emph{user} object represents an account in Asana that can be given access to various workspaces, projects, and tasks. \cr
}
}
\description{
Returns a compact representation of all of the parent goals of a goal.
}
\keyword{internal}
