% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{asn_get_portfolios}
\alias{asn_get_portfolios}
\title{Get multiple portfolios}
\usage{
asn_get_portfolios(workspace, limit, offset, opt_fields, owner)
}
\arguments{
\item{workspace}{(character scalar) The workspace or organization to filter portfolios on.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{owner}{(optional) (character scalar) The user who owns the portfolio. Currently, API users can only get a list of portfolios that they themselves own.}
}
\value{
A \emph{portfolio} gives a high-level overview of the status of multiple initiatives in Asana. Portfolios provide a dashboard overview of the state of multiple projects, including a progress report and the most recent \href{https://developers.asana.com/reference/project-statuses}{project status} update. Portfolios have some restrictions on size. Each portfolio has a max of 500 items and, like projects, a max of 20 custom fields.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab The name of the portfolio. \cr
}
}
\description{
Returns a list of the portfolios in compact representation that are owned by the current API user.
}
\keyword{internal}
