% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_get_project_memberships_for_project}
\alias{asn_get_project_memberships_for_project}
\title{Get memberships from a project}
\usage{
asn_get_project_memberships_for_project(
  project_gid,
  limit,
  offset,
  opt_fields,
  user
)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{user}{(optional) (character scalar) A string identifying a user. This can either be the string "me", an email, or the gid of a user.}
}
\value{
With the introduction of “comment-only” projects in Asana, a user’s membership in a project comes with associated permissions. These permissions (whether a user has full access to the project or comment-only access) are accessible through the project memberships endpoints described here.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   user \tab list \tab A \emph{user} object represents an account in Asana that can be given access to various workspaces, projects, and tasks. \cr
}
}
\description{
Returns the compact project membership records for the project.
}
\keyword{internal}
