% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_statuses.R
\name{asn_get_project_status}
\alias{asn_get_project_status}
\title{Get a project status}
\usage{
asn_get_project_status(project_status_gid, opt_fields)
}
\arguments{
\item{project_status_gid}{(character scalar) The project status update to get.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
\emph{Deprecated: new integrations should prefer the \verb{/status_updates/\{status_gid\}} route.}  Returns the complete record for a single status update.
}
\keyword{internal}
