% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_get_project_statuses_for_project}
\alias{asn_get_project_statuses_for_project}
\title{Get statuses from a project}
\usage{
asn_get_project_statuses_for_project(project_gid, limit, offset, opt_fields)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
\emph{Deprecated: new integrations should prefer the \code{status_update} resource.} A \emph{project status} is an update on the progress of a particular project, and is sent out to all project followers when created. These updates include both text describing the update and a color code intended to represent the overall state of the project: "green" for projects that are on track, "yellow" for projects at risk, and "red" for projects that are behind.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   title \tab character scalar \tab The title of the project status update. \cr
}
}
\description{
\emph{Deprecated: new integrations should prefer the \verb{/status_updates} route.}  Returns the compact project status update records for all updates on the project.
}
\keyword{internal}
