% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_templates.R
\name{asn_get_project_templates}
\alias{asn_get_project_templates}
\title{Get multiple project templates}
\usage{
asn_get_project_templates(limit, offset, opt_fields, team, workspace)
}
\arguments{
\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{team}{(optional) (character scalar) The team to filter projects on.}

\item{workspace}{(optional) (character scalar) The workspace to filter results on.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns the compact project template records for all project templates in the given team or workspace.
}
\keyword{internal}
