% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_get_section}
\alias{asn_get_section}
\title{Get a section}
\usage{
asn_get_section(section_gid, opt_fields)
}
\arguments{
\item{section_gid}{(character scalar) The globally unique identifier for the section.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns the complete record for a single section.
}
\keyword{internal}
