% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_get_tasks}
\alias{asn_get_tasks}
\title{Get multiple tasks}
\usage{
asn_get_tasks(
  assignee,
  completed_since,
  limit,
  modified_since,
  offset,
  opt_fields,
  project,
  section,
  workspace
)
}
\arguments{
\item{assignee}{(optional) (character scalar) The assignee to filter tasks on. If searching for unassigned tasks, assignee.any = null can be specified. \emph{Note: If you specify \code{assignee}, you must also specify the \code{workspace} to filter on.}}

\item{completed_since}{(optional) (datetime scalar) Only return tasks that are either incomplete or that have been completed since this time.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{modified_since}{(optional) (datetime scalar) Only return tasks that have been modified since the given time.  \emph{Note: A task is considered “modified” if any of its properties change, or associations between it and other objects are modified (e.g.  a task being added to a project). A task is not considered modified just because another object it is associated with (e.g. a subtask) is modified. Actions that count as modifying the task include assigning, renaming, completing, and adding stories.}}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{project}{(optional) (character scalar) The project to filter tasks on.}

\item{section}{(optional) (character scalar) The section to filter tasks on.}

\item{workspace}{(optional) (character scalar) The workspace to filter tasks on. \emph{Note: If you specify \code{workspace}, you must also specify the \code{assignee} to filter on.}}
}
\value{
The \emph{task} is the basic object around which many operations in Asana are centered.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab The name of the task. \cr
   resource_subtype \tab character scalar \tab The subtype of this resource. Different subtypes retain many of the same fields and behavior, but may render differently in Asana or represent resources with different semantic meaning. The resource_subtype \code{milestone} represent a single moment in time. This means tasks with this subtype cannot have a start_date. \cr
}
}
\description{
Returns the compact task records for some filtered set of tasks. Use one or more of the parameters provided to filter the tasks returned. You must specify a \code{project} or \code{tag} if you do not specify \code{assignee} and \code{workspace}.  For more complex task retrieval, use \href{https://developers.asana.com/reference/searchtasksforworkspace}{workspaces/{workspace_gid}/tasks/search}.
}
\keyword{internal}
