% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_task_lists.R
\name{asn_get_tasks_for_user_task_list}
\alias{asn_get_tasks_for_user_task_list}
\title{Get tasks from a user task list}
\usage{
asn_get_tasks_for_user_task_list(
  user_task_list_gid,
  completed_since,
  limit,
  offset,
  opt_fields
)
}
\arguments{
\item{user_task_list_gid}{(character scalar) Globally unique identifier for
the user task list.}

\item{completed_since}{(optional) (character scalar) Only return tasks that
are either incomplete or that have been completed since this time. Accepts
a date-time string or the keyword \emph{now}.}

\item{limit}{(optional) (integer scalar) Results per page. The number of
objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the
next page returned by the API. A pagination request will return an offset
token, which can be used as an input parameter to the next request. If an
offset is not passed in, the API will return the first page of results.
'Note: You can only pass in an offset that was returned to you via a
previously paginated request.'}

\item{opt_fields}{(optional) (character vector) Defines fields to return.
Some requests return \emph{compact} representations of objects in order to
conserve resources and complete the request more efficiently. Other times
requests return more information than you may need. This option allows you
to list the exact set of fields that the API should be sure to return for
the objects. The field names should be provided as paths, described below.
The id of included objects will always be returned, regardless of the field
options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns the compact list of tasks in a user’s My Tasks list. \emph{Note: Access
control is enforced for this endpoint as with all Asana API endpoints,
meaning a user’s private tasks will be filtered out if the API-authenticated
user does not have access to them.} \emph{Note: Both complete and incomplete tasks
are returned by default unless they are filtered out (for example, setting
\code{completed_since=now} will return only incomplete tasks, which is the default
view for “My Tasks” in Asana.)}
}
\keyword{internal}
