% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_memberships.R
\name{asn_get_team_membership}
\alias{asn_get_team_membership}
\title{Get a team membership}
\usage{
asn_get_team_membership(team_membership_gid, opt_fields)
}
\arguments{
\item{team_membership_gid}{(character scalar) NA}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns the complete team membership record for a single team membership.
}
\keyword{internal}
