% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_memberships.R
\name{asn_get_team_memberships}
\alias{asn_get_team_memberships}
\title{Get team memberships}
\usage{
asn_get_team_memberships(limit, offset, opt_fields, team, user, workspace)
}
\arguments{
\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{team}{(optional) (character scalar) Globally unique identifier for the team.}

\item{user}{(optional) (character scalar) A string identifying a user. This can either be the string "me", an email, or the gid of a user. This parameter must be used with the workspace parameter.}

\item{workspace}{(optional) (character scalar) Globally unique identifier for the workspace. This parameter must be used with the user parameter.}
}
\value{
This object represents a user's connection to a team.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   is_admin \tab logical scalar \tab Describes if the user is a team admin. \cr
   is_guest \tab logical scalar \tab Describes if the user is a guest in the team. \cr
   is_limited_access \tab logical scalar \tab Describes if the user has limited access to the team. \cr
   team \tab list \tab A \emph{team} is used to group related projects and people together within an organization. Each project in an organization is associated with a team. \cr
   user \tab list \tab A \emph{user} object represents an account in Asana that can be given access to various workspaces, projects, and tasks. \cr
}
}
\description{
Returns compact team membership records.
}
\keyword{internal}
