% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{asn_get_user}
\alias{asn_get_user}
\title{Get a user}
\usage{
asn_get_user(user_gid, opt_fields)
}
\arguments{
\item{user_gid}{(character scalar) A string identifying a user. This can either be the string "me", an email, or the gid of a user.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A \emph{user} object represents an account in Asana that can be given access to various workspaces, projects, and tasks.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab \emph{Read-only except when same user as requester}. The user’s name. \cr
   email \tab character scalar \tab The user's email address. \cr
   photo \tab list \tab A map of the user’s profile photo in various sizes, or null if no photo is set. Sizes provided are 21, 27, 36, 60, 128, and 1024. All images are in PNG format, except for 1024 (which is in JPEG format). \cr
   workspaces \tab list \tab Workspaces and organizations this user may access. Note\: The API will only return workspaces and organizations that also contain the authenticated user. \cr
}
}
\description{
Returns the full user record for the single user with the provided ID.
}
\keyword{internal}
