% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{asn_get_webhooks}
\alias{asn_get_webhooks}
\title{Get multiple webhooks}
\usage{
asn_get_webhooks(workspace, limit, offset, opt_fields, resource)
}
\arguments{
\item{workspace}{(character scalar) The workspace to query for webhooks in.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}

\item{resource}{(optional) (character scalar) Only return webhooks for the given resource.}
}
\value{
Webhook objects represent the state of an active subscription for a server to be updated with information from Asana. This schema represents the subscription itself, not the objects that are sent to the server. For information on those please refer to the \href{/reference/events}{event} schema.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   active \tab logical scalar \tab If true, the webhook will send events - if false it is considered inactive and will not generate events. \cr
   resource \tab list \tab A generic Asana Resource, containing a globally unique identifier. \cr
   target \tab character scalar \tab The URL to receive the HTTP POST. \cr
   created_at \tab datetime scalar \tab The time at which this resource was created. \cr
   filters \tab list \tab Whitelist of filters to apply to events from this webhook. If a webhook event passes any of the filters the event will be delivered; otherwise no event will be sent to the receiving server. \cr
   last_failure_at \tab datetime scalar \tab The timestamp when the webhook last received an error when sending an event to the target. \cr
   last_failure_content \tab character scalar \tab The contents of the last error response sent to the webhook when attempting to deliver events to the target. \cr
   last_success_at \tab datetime scalar \tab The timestamp when the webhook last successfully sent an event to the target. \cr
}
}
\description{
Get the compact representation of all webhooks your app has registered for the authenticated user in the given workspace.
}
\keyword{internal}
