% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_get_workspace}
\alias{asn_get_workspace}
\title{Get a workspace}
\usage{
asn_get_workspace(workspace_gid, opt_fields)
}
\arguments{
\item{workspace_gid}{(character scalar) Globally unique identifier for the workspace or organization.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A \emph{workspace} is the highest-level organizational unit in Asana. All projects and tasks have an associated workspace.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab The name of the workspace. \cr
   email_domains \tab list \tab The email domains that are associated with this workspace. \cr
   is_organization \tab logical scalar \tab Whether the workspace is an \emph{organization}. \cr
}
}
\description{
Returns the full workspace record for a single workspace.
}
\keyword{internal}
