% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_templates.R
\name{asn_instantiate_project}
\alias{asn_instantiate_project}
\title{Instantiate a project from a project template}
\usage{
asn_instantiate_project(project_template_gid, opt_fields)
}
\arguments{
\item{project_template_gid}{(character scalar) Globally unique identifier for the project template.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creates and returns a job that will asynchronously handle the project instantiation.  To form this request, it is recommended to first make a request to \href{/reference/getprojecttemplate}{get a project template}. Then, from the response, copy the \code{gid} from the object in the \code{requested_dates} array. This \code{gid} should be used in \code{requested_dates} to instantiate a project.  \emph{Note: The body of this request will differ if your workspace is an organization. To determine if your workspace is an organization, use the \href{/reference/workspaces}{is_organization} parameter.}
}
\keyword{internal}
