% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_project_save_as_template}
\alias{asn_project_save_as_template}
\title{Create a project template from a project}
\usage{
asn_project_save_as_template(project_gid, opt_fields)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Creates and returns a job that will asynchronously handle the project template creation. Note that while the resulting project template can be accessed with the API, it won't be visible in the Asana UI until Project Templates 2.0 is launched in the app. See more in \href{https://forum.asana.com/t/a-new-api-for-project-templates/156432}{this forum post}.
}
\keyword{internal}
