% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{asn_remove_custom_field_setting_for_portfolio}
\alias{asn_remove_custom_field_setting_for_portfolio}
\title{Remove a custom field from a portfolio}
\usage{
asn_remove_custom_field_setting_for_portfolio(portfolio_gid)
}
\arguments{
\item{portfolio_gid}{(character scalar) Globally unique identifier for the portfolio.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
Removes a custom field setting from a portfolio.
}
\keyword{internal}
