% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_remove_follower_for_task}
\alias{asn_remove_follower_for_task}
\title{Remove followers from a task}
\usage{
asn_remove_follower_for_task(task_gid, opt_fields)
}
\arguments{
\item{task_gid}{(character scalar) The task to operate on.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Removes each of the specified followers from the task if they are following. Returns the complete, updated record for the affected task.
}
\keyword{internal}
