% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolios.R
\name{asn_remove_item_for_portfolio}
\alias{asn_remove_item_for_portfolio}
\title{Remove a portfolio item}
\usage{
asn_remove_item_for_portfolio(portfolio_gid, opt_fields)
}
\arguments{
\item{portfolio_gid}{(character scalar) Globally unique identifier for the portfolio.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
Remove an item from a portfolio. Returns an empty data block.
}
\keyword{internal}
