% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_remove_project_for_task}
\alias{asn_remove_project_for_task}
\title{Remove a project from a task}
\usage{
asn_remove_project_for_task(task_gid, opt_fields)
}
\arguments{
\item{task_gid}{(character scalar) The task to operate on.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
Removes the task from the specified project. The task will still exist in the system, but it will not be in the project anymore.  Returns an empty data block.
}
\keyword{internal}
