% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_remove_user_for_team}
\alias{asn_remove_user_for_team}
\title{Remove a user from a team}
\usage{
asn_remove_user_for_team(team_gid, opt_fields)
}
\arguments{
\item{team_gid}{(character scalar) Globally unique identifier for the team.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
An empty object. Some endpoints do not return an object on success. The success is conveyed through a 2-- status code and returning an empty object.
}
\description{
The user making this call must be a member of the team in order to remove themselves or others.
}
\keyword{internal}
