% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_remove_user_for_workspace}
\alias{asn_remove_user_for_workspace}
\title{Remove a user from a workspace or organization}
\usage{
asn_remove_user_for_workspace(workspace_gid, opt_fields)
}
\arguments{
\item{workspace_gid}{(character scalar) Globally unique identifier for the
workspace or organization.}

\item{opt_fields}{(optional) (character vector) Defines fields to return.
Some requests return \emph{compact} representations of objects in order to
conserve resources and complete the request more efficiently. Other times
requests return more information than you may need. This option allows you
to list the exact set of fields that the API should be sure to return for
the objects. The field names should be provided as paths, described below.
The id of included objects will always be returned, regardless of the field
options.}
}
\value{
An empty object. Some endpoints do not return an object on success.
The success is conveyed through a 2-- status code and returning an empty
object.
}
\description{
Remove a user from a workspace or organization. The user making this call
must be an admin in the workspace. The user can be referenced by their
globally unique user ID or their email address. Returns an empty data record.
}
\keyword{internal}
