% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{asn_update_membership}
\alias{asn_update_membership}
\title{Update a membership}
\usage{
asn_update_membership(membership_gid, opt_fields)
}
\arguments{
\item{membership_gid}{(character scalar) Globally unique identifier for the membership.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
An existing membership can be updated by making a \code{PUT} request on the URL for that goal. Only the fields provided in the \code{data} block will be updated; any unspecified fields will remain unchanged. Memberships on \code{project}, \code{portfolio}, \code{team}, and \code{goals} can be updated.
}
\keyword{internal}
