% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_update_section}
\alias{asn_update_section}
\title{Update a section}
\usage{
asn_update_section(section_gid, opt_fields)
}
\arguments{
\item{section_gid}{(character scalar) The globally unique identifier for the section.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
A specific, existing section can be updated by making a PUT request on the URL for that project. Only the fields provided in the \code{data} block will be updated; any unspecified fields will remain unchanged. (note that at this time, the only field that can be updated is the \code{name} field.)  When using this method, it is best to specify only those fields you wish to change, or else you may overwrite changes made by another user since you last retrieved the task.  Returns the complete updated section record.
}
\keyword{internal}
