% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_update_story}
\alias{asn_update_story}
\title{Update a story}
\usage{
asn_update_story(story_gid, opt_fields)
}
\arguments{
\item{story_gid}{(character scalar) Globally unique identifier for the story.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Updates the story and returns the full record for the updated story. Only comment stories can have their text updated, and only comment stories and attachment stories can be pinned. Only one of \code{text} and \code{html_text} can be specified.
}
\keyword{internal}
