% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{asn_update_tag}
\alias{asn_update_tag}
\title{Update a tag}
\usage{
asn_update_tag(tag_gid, limit, offset, opt_fields)
}
\arguments{
\item{tag_gid}{(character scalar) Globally unique identifier for the tag.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A \emph{tag} is a label that can be attached to any task in Asana. It exists in a single workspace or organization.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab Name of the tag. This is generally a short sentence fragment that fits on a line in the UI for maximum readability. However, it can be longer. \cr
   color \tab character scalar \tab Color of the tag. \cr
   notes \tab character scalar \tab Free-form textual information associated with the tag (i.e. its description). \cr
   created_at \tab datetime scalar \tab The time at which this resource was created. \cr
   followers \tab list \tab Array of users following this tag. \cr
   permalink_url \tab character scalar \tab A url that points directly to the object within Asana. \cr
   workspace \tab list \tab A \emph{workspace} is the highest-level organizational unit in Asana. All projects and tasks have an associated workspace. \cr
}
}
\description{
Updates the properties of a tag. Only the fields provided in the \code{data} block will be updated; any unspecified fields will remain unchanged.  When using this method, it is best to specify only those fields you wish to change, or else you may overwrite changes made by another user since you last retrieved the tag.  Returns the complete updated tag record.
}
\keyword{internal}
