% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{asn_update_webhook}
\alias{asn_update_webhook}
\title{Update a webhook}
\usage{
asn_update_webhook(webhook_gid, opt_fields)
}
\arguments{
\item{webhook_gid}{(character scalar) Globally unique identifier for the webhook.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
An existing webhook's filters can be updated by making a PUT request on the URL for that webhook. Note that the webhook's previous \code{filters} array will be completely overwritten by the \code{filters} sent in the PUT request.
}
\keyword{internal}
