% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_update_workspace}
\alias{asn_update_workspace}
\title{Update a workspace}
\usage{
asn_update_workspace(workspace_gid, opt_fields)
}
\arguments{
\item{workspace_gid}{(character scalar) Globally unique identifier for the
workspace or organization.}

\item{opt_fields}{(optional) (character vector) Defines fields to return.
Some requests return \emph{compact} representations of objects in order to
conserve resources and complete the request more efficiently. Other times
requests return more information than you may need. This option allows you
to list the exact set of fields that the API should be sure to return for
the objects. The field names should be provided as paths, described below.
The id of included objects will always be returned, regardless of the field
options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
A specific, existing workspace can be updated by making a PUT request on the
URL for that workspace. Only the fields provided in the data block will be
updated; any unspecified fields will remain unchanged. Currently the only
field that can be modified for a workspace is its name. Returns the complete,
updated workspace record.
}
\keyword{internal}
