% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_header}
\alias{construct_header}
\title{Consstruct a header to send to the Asana API}
\usage{
construct_header(personal_access_token)
}
\arguments{
\item{personal_access_token}{Personal Access Token. Available in your Asana account}
}
\description{
Consstruct a header to send to the Asana API
}
\author{
Ryan Sullivan
}
