% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.call_api}
\alias{.call_api}
\title{Call the Asana API}
\usage{
.call_api(endpoint, query, body, method, token = Sys.getenv("ASN_API_TOKEN"))
}
\arguments{
\item{endpoint}{The path to an endpoint. Optionally, a list with the path
plus variables to \code{\link[glue:glue]{glue::glue()}} into the path.}

\item{query}{An optional list of parameters to pass in the query portion of
the request.}

\item{body}{An optional list of parameters to pass in the body portion of the
request. Should have additional class "json" or "multipart". See the
associated http2 functions for details.}

\item{method}{If the method is something other than GET or POST, supply it.
Case is ignored.}

\item{token}{Your API token. Probably always pass this in the environment
variable.}
}
\value{
The result of the call.
}
\description{
This is unexported because I intend to use it from endpoint-specific
functions.
}
\keyword{internal}
