% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.new_asn_response}
\alias{.new_asn_response}
\title{Construct an asn_response}
\usage{
.new_asn_response(response_data)
}
\arguments{
\item{response_data}{The \code{data} portion of a response from
\code{httr2::resp_body_json()}.}
}
\value{
An object with additional class "asn_response", or \code{NULL}.
}
\description{
Construct an asn_response
}
\keyword{internal}
