% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.parse_error}
\alias{.parse_error}
\title{Extract error information from response}
\usage{
.parse_error(response)
}
\arguments{
\item{response}{A response that failed.}
}
\value{
A named character vector of error information.
}
\description{
Extract error information from response
}
\keyword{internal}
