% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_requests.R
\name{.parse_response}
\alias{.parse_response}
\title{Parse the returned response}
\usage{
.parse_response(response)
}
\arguments{
\item{response}{A raw response returned from the Asana api.}
}
\value{
Almost always a list.
}
\description{
Parse the returned response
}
\keyword{internal}
