% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workspace.R
\name{get_workspace}
\alias{get_workspace}
\title{Return the workspace specified by the Id.}
\usage{
get_workspace(id = NULL)
}
\arguments{
\item{id}{Globally unique ID of the workspace.}
}
\description{
Return the workspace specified by the Id.
}
\examples{
\dontrun{
#A workspace is the highest-level organizational unit in Asana. All projects and tasks have an associated workspace.
get_workspace(id = 1234) )
}
}
\author{
Ryan Sullivan
}
