% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{registerOptions}
\alias{registerOptions}
\title{Register Asana Personal Access Token}
\usage{
registerOptions(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{Logical. If TRUE, verbose messages will be printed to the R console. Defaults to TRUE.}

\item{...}{A 'personal_access_token' to register with Asana Connect. (see example). See also \code{\link{asanaRconfigFile}} for an explanation of the personal_access_token parameters.}
}
\description{
This function registers the user's Asana Personal Access Token.
}
\examples{
\dontrun{
#Register your Asana Personal Access Token if you haven't already
#Note that you need to pass the 'personal_access_token'
#parameter if you call this function for the first time.
registerOptions(personal_access_token="<YOUR-PA-TOKEN>" )
} 
}
\seealso{
See \url{https://asana.com/developers/documentation/getting-started/auth#personal-access-token}
}
\author{
Ryan Sullivan
}
