% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_add_custom_field_setting_for_project}
\alias{asn_add_custom_field_setting_for_project}
\title{Add a custom field to a project}
\usage{
asn_add_custom_field_setting_for_project(project_gid)
}
\arguments{
\item{project_gid}{(character scalar) Globally unique identifier for the project.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Custom fields are associated with projects by way of custom field settings.  This method creates a setting for the project.
}
\keyword{internal}
