% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{asn_create_batch_request}
\alias{asn_create_batch_request}
\title{Submit parallel requests}
\usage{
asn_create_batch_request(opt_fields)
}
\arguments{
\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A response object returned from a batch request.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   body \tab list \tab The JSON body that the invoked endpoint returned. \cr
   headers \tab list \tab A map of HTTP headers specific to this result. This is primarily used for returning a \code{Location} header to accompany a \verb{201 Created} result.  The parent HTTP response will contain all common headers. \cr
   status_code \tab integer scalar \tab The HTTP status code that the invoked endpoint returned. \cr
}
}
\description{
Make multiple requests in parallel to Asana's API.
}
\keyword{internal}
