% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{asn_get_attachment}
\alias{asn_get_attachment}
\title{Get an attachment}
\usage{
asn_get_attachment(attachment_gid, opt_fields)
}
\arguments{
\item{attachment_gid}{(character scalar) Globally unique identifier for the
attachment.}

\item{opt_fields}{(optional) (character vector) Defines fields to return.
Some requests return \emph{compact} representations of objects in order to
conserve resources and complete the request more efficiently. Other times
requests return more information than you may need. This option allows you
to list the exact set of fields that the API should be sure to return for
the objects. The field names should be provided as paths, described below.
The id of included objects will always be returned, regardless of the field
options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Get the full record for a single attachment.
}
\keyword{internal}
