% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_get_section}
\alias{asn_get_section}
\title{Get a section}
\usage{
asn_get_section(section_gid, opt_fields)
}
\arguments{
\item{section_gid}{(character scalar) The globally unique identifier for the section.}

\item{opt_fields}{(optional) (list) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A \emph{section} is a subdivision of a project that groups tasks together. It can either be a header above a list of tasks in a list view or a column in a board view of a project.\tabular{lll}{
   \strong{Property} \tab \strong{Class} \tab \strong{Description} \cr
   gid \tab character scalar \tab Globally unique identifier of the resource, as a string. \cr
   resource_type \tab character scalar \tab The base type of this resource. \cr
   name \tab character scalar \tab The name of the section (i.e. the text displayed as the section header). \cr
   created_at \tab datetime scalar \tab The time at which this resource was created. \cr
   project \tab list \tab A \emph{project} represents a prioritized list of tasks in Asana or a board with columns of tasks represented as cards. It exists in a single workspace or organization and is accessible to a subset of users in that workspace or organization, depending on its permissions. \cr
   projects \tab list \tab \emph{Deprecated - please use project instead} \cr
}
}
\description{
Returns the complete record for a single section.
}
\keyword{internal}
