% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_updates.R
\name{asn_get_statuses_for_object}
\alias{asn_get_statuses_for_object}
\title{Get status updates from an object}
\usage{
asn_get_statuses_for_object(parent, created_since, limit, offset, opt_fields)
}
\arguments{
\item{parent}{(character scalar) Globally unique identifier for object to fetch statuses from. Must be a GID for a project, portfolio, or goal.}

\item{created_since}{(optional) (datetime scalar) Only return statuses that have been created since the given time.}

\item{limit}{(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.}

\item{offset}{(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns the compact status update records for all updates on the object.
}
\keyword{internal}
