% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{asn_get_user_task_list_for_user}
\alias{asn_get_user_task_list_for_user}
\title{Get a user's task list}
\usage{
asn_get_user_task_list_for_user(user_gid, workspace, opt_fields)
}
\arguments{
\item{user_gid}{(character scalar) A string identifying a user. This can either be the string "me", an email, or the gid of a user.}

\item{workspace}{(character scalar) The workspace in which to get the user task list.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Returns the full record for a user's task list.
}
\keyword{internal}
