% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_update_custom_field}
\alias{asn_update_custom_field}
\title{Update a custom field}
\usage{
asn_update_custom_field(custom_field_gid, opt_fields)
}
\arguments{
\item{custom_field_gid}{(character scalar) Globally unique identifier for the custom field.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
A specific, existing custom field can be updated by making a PUT request on the URL for that custom field. Only the fields provided in the \code{data} block will be updated; any unspecified fields will remain unchanged When using this method, it is best to specify only those fields you wish to change, or else you may overwrite changes made by another user since you last retrieved the custom field. A custom field’s \code{type} cannot be updated. An enum custom field’s \code{enum_options} cannot be updated with this endpoint. Instead see “Work With Enum Options” for information on how to update \code{enum_options}. Locked custom fields can only be updated by the user who locked the field. Returns the complete updated custom field record.
}
\keyword{internal}
