% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_options.R
\name{asn_update_enum_option}
\alias{asn_update_enum_option}
\title{Update an enum option}
\usage{
asn_update_enum_option(enum_option_gid, opt_fields)
}
\arguments{
\item{enum_option_gid}{(character scalar) Globally unique identifier for the enum option.}

\item{opt_fields}{(optional) (character vector) Defines fields to return. Some requests return \emph{compact} representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.}
}
\value{
A generic Asana Resource, containing a globally unique identifier.
}
\description{
Updates an existing enum option. Enum custom fields require at least one enabled enum option. Locked custom fields can only be updated by the user who locked the field. Returns the full record of the updated enum option.
}
\keyword{internal}
