% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_header}
\alias{construct_header}
\title{Construct a header to send to qualtrics API}
\usage{
construct_header(personal_access_token)
}
\arguments{
\item{personal_access_token}{Personal Access Token. Available in your Asana account.}
}
\description{
Construct a header to send to qualtrics API
}
