% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workspaces.R
\name{get_workspaces}
\alias{get_workspaces}
\title{Get workspaces from your Asana account}
\usage{
get_workspaces(limit = 10, all_pages = TRUE)
}
\arguments{
\item{limit}{Test}

\item{all_pages}{Test}
}
\description{
Get workspaces from your Asana account
}
\examples{
\dontrun{
#Use this function 
#A workspace is the highest-level organizational unit in Asana. All projects and tasks have an associated workspace.
 
get_workspaces(limit = 10, all_pages = TRUE )
}
}
